Refactor API exception handling and configuration loading***
